+++ tests/test_testutil.py
@@ -20,9 +20,9 @@
 import unittest
 
 from repoze.what.adapters.testutil import GroupsAdapterTester, \
-                                                  PermissionsAdapterTester
+                                          PermissionsAdapterTester
 
-from base import FakeGroupSourceAdapter, FakePermissionSourceAdapter
+from repoze.what.testutil import FakeGroupSourceAdapter, FakePermissionSourceAdapter
 
 class TestGroupsAdapterTester(GroupsAdapterTester, unittest.TestCase):
     def setUp(self):